# optimise `secp256k1` on osx