`new`
|            Method |      Mean |     Error |    StdDev |
|------------------ |----------:|----------:|----------:|
|   VerifySignature |  1.553 ms | 0.0138 ms | 0.0129 ms |

`old`
|            Method |      Mean |     Error |    StdDev |
|------------------ |----------:|----------:|----------:|
|   VerifySignature |  1.649 ms | 0.0327 ms | 0.0726 ms |